src/functions/transcription_processor/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-translate==3.*
google-cloud-aiplatform==1.*
openai==1.*
structlog==24.1.0
srt==3.*
SQLAlchemy==2.0.41
psycopg2-binary==2.9.*
mecab-python3==1.0.*
unidic-lite==1.0.8